types file and linting fixes
