--- packages/redux-slice-factory/src/slice-factories/meta-slice/index.ts
@@ -1,1 +1,0 @@
-export type { IMetaSliceReducers } from './MetaSlice';
